Issues with Android Permissions when manipulating DB file

+++ ViewModels/SettingsPageViewModel.cs
@@ -10,8 +10,10 @@ namespace FarmOrganizer.ViewModels
 {
     public partial class SettingsPageViewModel : ObservableObject
     {
+        #region Preference Keys
         public const string AppThemeKey = "appTheme";
         public const string LedgerPage_DefaultCropField = "ledger_defaultCropFieldKey";
+        #endregion
 
         [ObservableProperty]
         private List<string> appThemes;
@@ -24,6 +26,8 @@ namespace FarmOrganizer.ViewModels
         private CropField defaultCropField;
         [ObservableProperty]
         private bool cropFieldPickerEnabled = true;
+
+        private const string _IOAlertNoPermissions = "Aby aplikacja mogła zresetować bazę danych, potrzebne są odpowiednie uprawnienia.\nJeżeli ten komunikat pokazuje się po zrestartowaniu aplikacji, możliwe że wymagane jest zresetowanie odmówionych uprawnień. Przejdź do ustawień swojego telefonu, a następnie w sekcji 'Aplikacje', odnajdź FarmOrganizer i nadaj mu uprawnienia do zapisu i odczytu plików.";
 
         public SettingsPageViewModel()
         {
@@ -81,8 +85,13 @@ namespace FarmOrganizer.ViewModels
                 "Tej akcji nie można odwrócić. Czy jesteś pewny aby kontynuować?",
                 "Tak", "Nie"))
             {
+                if (!await DatabaseFile.RequestPermissions())
+                {
+                    App.AlertSvc.ShowAlert("Błąd", _IOAlertNoPermissions);
+                    return;
+                }
                 await DatabaseFile.Delete();
-                await MainThread.InvokeOnMainThreadAsync(DatabaseFile.Create);
+                await DatabaseFile.Create();
             }
         }
 
@@ -91,6 +100,11 @@ namespace FarmOrganizer.ViewModels
         {
             try
             {
+                if (!await DatabaseFile.RequestPermissions())
+                {
+                    App.AlertSvc.ShowAlert("Błąd", _IOAlertNoPermissions);
+                    return;
+                }
                 FolderPickerResult folder = await FolderPicker.PickAsync(default);
                 if (!folder.IsSuccessful)
                     return;
@@ -111,6 +125,11 @@ namespace FarmOrganizer.ViewModels
         {
             try
             {
+                if (!await DatabaseFile.RequestPermissions())
+                {
+                    App.AlertSvc.ShowAlert("Błąd", _IOAlertNoPermissions);
+                    return;
+                }
                 await DatabaseFile.CreateBackup();
                 FileResult file = await FilePicker.PickAsync();
                 if (file == null)
